\name{datafrequency}
\alias{datafrequency}
\title{Calculate the frequency}
\usage{
datafrequency(data = NULL, unit = NULL, verbose = FALSE)
}
\arguments{
\item{data}{a time index vector}
\item{unit}{unit of time index}
\item{verbose}{if TRUE print information about progress}
}
\description{
Calculate the frequency of data based on a time index vector.
}
