% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\alias{retrieve.ncdf4}
\alias{retrieve.rcm}
\title{Retrieve field data from a netcdf file.}
\usage{
retrieve(ncfile,param="auto",path=NULL,verbose=FALSE,...)
}
\arguments{
\item{ncfile}{A character string of a netcdf file name with or without the full path or any object of class 'ncdf4'.}
\item{path}{A character string with the path to the file \code{ncfile}}
\item{param}{Parameter or element type. There are several core parameters or
elements as well as a number of additional parameters. The parameters or
elements are: auto = automatic selection.  precip, prcp, pr = Precipitation
(mm) tas, tavg = 2m-surface temperature (in degrees Celcius) tmax, tasmax =
Maximum temperature (in degrees Celcius) tmin, tasmin = Minimum temperature
(in degrees Celcius)}
\item{verbose}{Logical value defaulting to FALSE. If FALSE, do not display
comments (silent mode). If TRUE, displays extra information on progress.}
\item{\dots}{Additional input.}
}
\value{
A "zoo" object with additional attributes.
}
\description{
Retrieve data from a netcdf file and return a zoo field object with
attributes.  \code{retrieve} assumes data on a regular lon-lat grid and
\code{retrieve.rcm} reads data on irregular (rotated) grid, typically output
from RCMs.
}